requirements.txt
google-generativeai
selenium
discord.py
pillow
python-dotenv
flask
requests
